testing/web-platform/mozilla/tests/indic-detection/baskar-jagran.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
×¢»Ü »ëã âõÚ×¢ÇÜ ×ğ¢ âêØü âğ ¿õÍæ »ıã ãñĞ §â·ğUUU ÌÜ ·UUUè ¥æÖæ ÚçQ¤× ãñ, çÁâ ßÁã âğ §âğ "ÜæÜ »ıã" ·ğUUU Ùæ× âğ Öè ÁæÙæ ÁæÌæ ãñĞ âõÚ×¢ÇÜ ·ğUUU »ıã Îô ÌÚã ·ğUUU ãôÌğ ãñ¢ - "SÍÜèØ »ıã" çÁÙ·UUUæ ÌÜ ¥æÖæâèØ ãôÌæ ãñ ¥æñÚ
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://hi.wikipedia.org/w/index.php?title=%E0%A4%AE%E0%A4%82%E0%A4%97%E0%A4%B2_%E0%A4%97%E0%A5%8D%E0%A4%B0%E0%A4%B9&oldid=5105576">&#2350;&#2306;&#2327;&#2354; &#2327;&#2381;&#2352;&#2361;</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
